Split process module
